Renamed Timer.tsx to TimerComponent.tsx to avoid confusion
